+++ xcube/webapi/handlers.py
@@ -21,25 +21,33 @@
 
 import datetime
 import json
+import logging
+import os.path
+import pathlib
 
 from tornado.ioloop import IOLoop
 
-from . import __version__, __description__
 from .controllers.catalogue import get_datasets, get_dataset_coordinates, get_color_bars, get_dataset
 from .controllers.places import find_places, find_dataset_places
 from .controllers.tiles import get_dataset_tile, get_dataset_tile_grid, get_ne2_tile, get_ne2_tile_grid, get_legend
 from .controllers.time_series import get_time_series_info, get_time_series_for_point, get_time_series_for_geometry, \
     get_time_series_for_geometry_collection, get_time_series_for_feature_collection
 from .controllers.wmts import get_wmts_capabilities_xml
+from .defaults import SERVER_NAME, SERVER_DESCRIPTION
 from .errors import ServiceBadRequestError
+from .s3util import dict_to_xml, list_s3_bucket_v1, list_bucket_result_to_xml, list_s3_bucket_v2, \
+    mtime_to_str, str_to_etag
 from .service import ServiceRequestHandler
 from ..util.timecoord import timestamp_to_iso_string
+from ..version import version
 
 __author__ = "Norman Fomferra (Brockmann Consult GmbH)"
 
 _WMTS_VERSION = "1.0.0"
 _WMTS_TILE_FORMAT = "image/png"
+_LOG = logging.getLogger('xcube')
 
+_LOG_S3BUCKET_HANDLER = False
 
 # noinspection PyAbstractClass
 class WMTSKvpHandler(ServiceRequestHandler):
@@ -119,6 +127,7 @@ class GetDatasetsHandler(ServiceRequestHandler):
         self.write(json.dumps(response, indent=None if details else 2))
 
 
+# noinspection PyAbstractClass
 class GetDatasetHandler(ServiceRequestHandler):
 
     def get(self, ds_id: str):
@@ -129,6 +138,118 @@ class GetDatasetHandler(ServiceRequestHandler):
 
 
 # noinspection PyAbstractClass
+class ListS3BucketHandler(ServiceRequestHandler):
+
+    async def get(self):
+
+        prefix = self.get_query_argument('prefix', default=None)
+        delimiter = self.get_query_argument('delimiter', default=None)
+        max_keys = int(self.get_query_argument('max-keys', default='1000'))
+        list_s3_bucket_params = dict(prefix=prefix, delimiter=delimiter,
+                                     max_keys=max_keys)
+
+        list_type = self.get_query_argument('list-type', default=None)
+        if list_type is None:
+            marker = self.get_query_argument('marker', default=None)
+            list_s3_bucket_params.update(marker=marker)
+            list_s3_bucket = list_s3_bucket_v1
+        elif list_type == '2':
+            start_after = self.get_query_argument('start-after', default=None)
+            continuation_token = self.get_query_argument('continuation-token', default=None)
+            list_s3_bucket_params.update(start_after=start_after, continuation_token=continuation_token)
+            list_s3_bucket = list_s3_bucket_v2
+        else:
+            raise ServiceBadRequestError(f'Unknown bucket list type {list_type!r}')
+
+        if _LOG_S3BUCKET_HANDLER:
+            _LOG.info(f'GET: list_s3_bucket_params={list_s3_bucket_params}')
+        bucket_mapping = self.service_context.get_s3_bucket_mapping()
+        list_bucket_result = list_s3_bucket(bucket_mapping, **list_s3_bucket_params)
+        if _LOG_S3BUCKET_HANDLER:
+            import json
+            _LOG.info(f'-->\n{json.dumps(list_bucket_result, indent=2)}')
+
+        xml = list_bucket_result_to_xml(list_bucket_result)
+        self.set_header('Content-Type', 'application/xml')
+        self.write(xml)
+        await self.flush()
+
+
+# noinspection PyAbstractClass
+class GetS3BucketObjectHandler(ServiceRequestHandler):
+    async def head(self, ds_id: str, path: str = ''):
+        key, local_path = self._get_key_and_local_path(ds_id, path)
+        if _LOG_S3BUCKET_HANDLER:
+            _LOG.info(f'HEAD: key={key!r}, local_path={local_path!r}')
+        if local_path is None or not local_path.exists():
+            await self._key_not_found(key)
+            return
+        self.set_header('ETag', str_to_etag(str(local_path)))
+        self.set_header('Last-Modified', mtime_to_str(local_path.stat().st_mtime))
+        if local_path.is_file():
+            self.set_header('Content-Length', local_path.stat().st_size)
+        else:
+            self.set_header('Content-Length', 0)
+        await self.finish()
+
+    async def get(self, ds_id: str, path: str = ''):
+        key, local_path = self._get_key_and_local_path(ds_id, path)
+        if _LOG_S3BUCKET_HANDLER:
+            _LOG.info(f'GET: key={key!r}, local_path={local_path!r}')
+        if local_path is None or not local_path.exists():
+            await self._key_not_found(key)
+            return
+        self.set_header('ETag', str_to_etag(str(local_path)))
+        self.set_header('Last-Modified', mtime_to_str(local_path.stat().st_mtime))
+        self.set_header('Content-Type', 'binary/octet-stream')
+        if local_path.is_file():
+            self.set_header('Content-Length', local_path.stat().st_size)
+            chunk_size = 1024 * 1024
+            with open(str(local_path), 'rb') as fp:
+                while True:
+                    chunk = fp.read(chunk_size)
+                    if len(chunk) == 0:
+                        break
+                    self.write(chunk)
+                    await self.flush()
+        else:
+            self.set_header('Content-Length', 0)
+            await self.finish()
+
+    def _key_not_found(self, key: str):
+        self.set_header('Content-Type', 'application/xml')
+        self.set_status(404)
+        return self.finish(dict_to_xml('Error',
+                                       dict(Code='NoSuchKey',
+                                            Message='The specified key does not exist.',
+                                            Key=key)))
+
+    def _get_key_and_local_path(self, ds_id: str, path: str):
+        descriptor = self.service_context.get_dataset_descriptor(ds_id)
+        file_system = descriptor.get('FileSystem', 'local')
+        required_file_system = 'local'
+        if file_system != required_file_system:
+            raise ServiceBadRequestError(f'AWS S3 data access: currently, only datasets in'
+                                         f' file system {required_file_system!r} are supported,'
+                                         f' but dataset {ds_id!r} uses file system {file_system!r}')
+
+        key = f'{ds_id}/{path}'
+
+        # validate path
+        if path and '..' in path.split('/'):
+            raise ServiceBadRequestError(f'AWS S3 data access: received illegal key {key!r}')
+
+        local_path = descriptor.get('Path')
+        if os.path.isabs(local_path):
+            local_path = os.path.join(local_path, path)
+        else:
+            local_path = os.path.join(self.service_context.base_dir, local_path, path)
+
+        local_path = os.path.normpath(local_path)
+
+        return key, pathlib.Path(local_path)
+
+
 class GetDatasetCoordsHandler(ServiceRequestHandler):
 
     def get(self, ds_id: str, dim_name: str):
@@ -299,9 +420,9 @@ class InfoHandler(ServiceRequestHandler):
                                               freq="ms")
         server_time = timestamp_to_iso_string(datetime.datetime.now(), freq="ms")
         self.set_header('Content-Type', 'application/json')
-        self.write(json.dumps(dict(name='xcube_server',
-                                   description=__description__,
-                                   version=__version__,
+        self.write(json.dumps(dict(name=SERVER_NAME,
+                                   description=SERVER_DESCRIPTION,
+                                   version=version,
                                    configTime=config_time,
                                    serverTime=server_time),
                               indent=2))