+++ CHANGES.md
@@ -2,6 +2,12 @@
 
 ### Enhancements
 
+* `xcube serve` does now offer a AWS S3 compatible data access API (#115):
+   - List bucket objects: `/s3bucket`, see AWS 
+     docs [GET](https://docs.aws.amazon.com/AmazonS3/latest/API/v2-RESTBucketGET.html)
+   - Get bucket object: `/s3bucket/{ds_id}/{path}`, 
+     see AWS docs [HEAD](https://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectHEAD.html) 
+     and [GET](https://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html)
 * `xcube serve` now verifies that a configured cube is valid once it is opened. (#107)
 * Added new CLI command `xcube verify` performing data cube verification. (#19)
 * Reworked `xcube extract` to be finally useful and effective for point data extraction. (#102) 
@@ -9,7 +15,6 @@
 * `xcube server`can now limit time series to a maximum number of 
   valid (not NaN) values. To activate, pass optional query parameter `maxValids` to the various `/ts`
   functions. The special value `-1` will restrict the result to contain only valid values. (#113) 
-* `xcube server`can now filter datasets by point coordinate, e.g. `/dataset?point=12.5,52.8`. (#50) 
 * Reworked `xcube gen` to be more user-friendly and more consistent with other tools. 
   The changes are
   - Removed `--dir` and `--name` options and replaced it by single `--output` option,